Rename empty case event
