+++ src/main/java/uk/gov/hmcts/divorce/systemupdate/schedule/bulkaction/FailedBulkCaseRemover.java
@@ -16,7 +16,7 @@ import java.util.List;
 import java.util.function.Predicate;
 
 import static org.springframework.util.CollectionUtils.isEmpty;
-import static uk.gov.hmcts.divorce.bulkaction.ccd.event.SystemDropEmptyCase.SYSTEM_EMPTY_CASE;
+import static uk.gov.hmcts.divorce.bulkaction.ccd.event.SystemEmptyCase.SYSTEM_EMPTY_CASE;
 import static uk.gov.hmcts.divorce.bulkaction.ccd.event.SystemRemoveFailedCases.SYSTEM_REMOVE_FAILED_CASES;
 
 @Component
